Mieux

+++ Main_robot/controleur.h
@@ -14,7 +14,7 @@ struct MotorPWM {
 extern float integral;
 extern int32_t last_T;
 
-MotorPWM controleur(EncoderData data, int8_t linePosition, float gains[4], int32_t psi_ref);
+MotorPWM controleur(EncoderData data, int32_t linePosition, float gains[4], int32_t psi_ref);
 float integrale(float lambda);
 
 #endif
